imegrate to color micro service
